rename peddy parsing module
